Refactor apply code into its own package
